requirements.txt
Pillow~=10.4.0
Flask~=3.0.3
tensorflow>=2.0.0
numpy~=1.26.4